Fixed wrong type

+++ Assets/Scripts/AnimatorStates/HurtState.cs
@@ -6,7 +6,7 @@ public class HurtState : StateMachineBehaviour
 {
     override public void OnStateEnter(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
     {
-        if (animator.GetFloat("Health") <= 0)
+        if (animator.GetInteger("Health") <= 0)
         {
             animator.SetBool("IsDying", true);
